WIP: Start refactoring SnakeObj class and prepare Map class
