+++ SnakeController/Map.hpp
@@ -1,0 +1,18 @@
+//
+// Created by a19stude on 4/18/19.
+//
+
+#ifndef REFACTORINGWORKSHOP_MAP_HPP
+#define REFACTORINGWORKSHOP_MAP_HPP
+
+
+class Map {
+
+private:
+	m_mapDimension = std::make_pair(width, height);
+	m_foodPosition = std::make_pair(foodX, foodY);
+
+};
+
+
+#endif //REFACTORINGWORKSHOP_MAP_HPP
