+++ SnakeController/SnakeController.hpp
@@ -6,17 +6,13 @@
 
 #include "IEventHandler.hpp"
 #include "SnakeInterface.hpp"
+#include "Snake.hpp"
 
 class Event;
 class IPort;
 
 namespace Snake
 {
-struct ConfigurationError : std::logic_error
-{
-    ConfigurationError();
-};
-
 struct UnexpectedEventException : std::runtime_error
 {
     UnexpectedEventException();
@@ -40,14 +36,8 @@ private:
     std::pair<int, int> m_mapDimension;
     std::pair<int, int> m_foodPosition;
 
-    struct Segment
-    {
-        int x;
-        int y;
-    };
+    SnakeObj snake;
 
-    std::list<Segment> m_segments;
-    Direction m_currentDirection;
 
     void handleTimeoutInd();
     void handleDirectionInd(std::unique_ptr<Event>);
@@ -55,12 +45,7 @@ private:
     void handleFoodResp(std::unique_ptr<Event>);
     void handlePauseInd(std::unique_ptr<Event>);
 
-    bool isSegmentAtPosition(int x, int y) const;
-    Segment calculateNewHead() const;
-    void updateSegmentsIfSuccessfullMove(Segment const& newHead);
-    void addHeadSegment(Segment const& newHead);
-    void removeTailSegmentIfNotScored(Segment const& newHead);
-    void removeTailSegment();
+
 
     bool isPositionOutsideMap(int x, int y) const;
